Make GD singleton friendly to java
